Application/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
sklearn==0.0
torch==1.5.0
torchvision==0.6.0
Werkzeug==1.0.1
